backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
colorama==0.4.6
exceptiongroup==1.3.1
fastapi==0.136.1
greenlet==3.5.0
h11==0.16.0
idna==3.13
pydantic==2.13.3
pydantic_core==2.46.3
psycopg2-binary==2.9.9
python-dotenv==1.2.2
setuptools==65.5.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
email-validator==2.1.1